+++ core.webui/Models/CartModel.cs
@@ -8,7 +8,7 @@ namespace core.webui.Models
         public int CartId { get; set; }
         public List<CartItemModel> CartItems { get; set; }
 
-        public double TotalPrice()
+        public decimal TotalPrice()
         {
             return CartItems.Sum(i=>i.Price*i.Quantity);
         }
@@ -19,7 +19,7 @@ namespace core.webui.Models
         public int CartItemId { get; set; }
         public int ProductId { get; set; }
         public string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string ImageUrl { get; set; }
         public int Quantity { get; set; }
     }
